[tmux] remove unused
Some things do not work in tmux or tmux>neovim when you pretend to be an
older terminal. Specifically curly underlines. the only place where you
may still want to pretend to be an older terminal is when ssh'ing to a
server that may not support modern $TERM such as xterm-ghostty